move logger to separate file
